# By Andrey Derkach (98) and others
# Via Viktor Kleschenko (38) and others
* commit '9fbb2ff3e6b17a7400bb726e85318d94689119ff [formerly f99d2d548eaf19dd2ed9b5d6b46df93c6a943816]': (542 commits)
  [DF-6970] Suppress SetVariant to self.
  DF-5346 make correct cache cleanup in RenderTechniqueSingleton::ClearRenderTechniques
  DF-6590 fix DF-6941, DF-6940, DF-6939, DF-6929 fix iOS bug see comments + fix comment
  DF-6590 fix DF-6941, DF-6940, DF-6939, DF-6929 fix iOS bug see comments
  [DF-6937] FileSystem::IsDirectory fix
  DF-5346 prettify code
  [DF-6618] Change GetTickCount() Android implementation to use clock_gettime( CLOCK_MONOTONIC, &ts )
  DF-6590 fix review comments
  DF-5346 refine RenderTechniqueSingleton::ClearRenderTechniques()
  [DF-6792] We don't care about viewport parent because there are many possibilities to make same error at animation works.
  DF-6590 fix DF-6929 old value stay in text fields, old problem with YES / NO during render into texture
  DF-5346 add cache cleanup methods
  Fix macro SAFE_DELETE and SAFE_DELETE_ARRAY
  [DF-5824] Add fix to Nexus 4
  [DF-6926] Landscape draw fix
  DF-5943 convert typedef to C++11 using
  [DF-6925] [DF-6928] Fixed problem with resource folders
  Development build fix
  DF-6744 open atc crashfix
  [DF-185] Fixed show / hide cursor in windows
  ...


Former-commit-id: 4e3f19af8fa9d1e790dc9fba00891bfd73a09d90

+++ Sources/Internal/Platform/Qt5/QtLayer.cpp
@@ -27,7 +27,6 @@
 =====================================================================================*/
 
 
-
 #include "Platform/Qt5/QtLayer.h"
 
 #include "Render/RenderManager.h"
@@ -165,7 +164,7 @@ void QtLayer::KeyPressed(char16 key, int32 count, uint64 timestamp)
     UIEvent ev;
     ev.keyChar = 0;
     ev.phase = UIEvent::PHASE_KEYCHAR;
-    ev.timestamp = timestamp;
+    ev.timestamp = static_cast<float64>(timestamp);
     ev.tapCount = 1;
     ev.tid = key;
